move db initialization to a separate class
